# apps/backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/branches/{branchId}/kitchen-display": {
      "get": {
        "operationId": "DisplayController_getKitchenDisplay",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/DisplayItemResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get pending/preparing items for kitchen",
        "tags": [
          "Kitchen/Bar Display"
        ]
      }
    },
    "/branches/{branchId}/bar-display": {
      "get": {
        "operationId": "DisplayController_getBarDisplay",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/DisplayItemResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get pending/preparing items for bar",
        "tags": [
          "Kitchen/Bar Display"
        ]
      }
    },
    "/display-items/{id}/status": {
      "patch": {
        "operationId": "DisplayController_updateItemStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDisplayItemStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DisplayItemResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mark item preparing/ready/unavailable",
        "tags": [
          "Kitchen/Bar Display"
        ]
      }
    },
...
  },
  "info": {
    "title": "milky",
    "description": "The milky API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "MetaDataDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "total": {
            "type": "number",
            "example": 50
          }
        },
        "required": [
          "page",
          "limit",
          "total"
        ]
      },
      "GenericResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaDataDto"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "DisplayItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "productNameSnapshot": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "stationType": {
            "type": "string",
            "enum": [
              "KITCHEN",
              "BAR"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PREPARING",
              "READY",
              "SERVED",
              "UNAVAILABLE",
              "CANCELLED"
            ]
          }
        },
        "required": [
          "id",
          "orderId",
          "productNameSnapshot",
          "quantity",
          "stationType",
          "status"
        ]
      },
      "UpdateDisplayItemStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PREPARING",
              "READY",
              "SERVED",
              "UNAVAILABLE",
              "CANCELLED"
            ]
          },
          "unavailableReason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}